backend/requirements.txt
aiosqlite==0.22.1
alembic==1.17.1
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
asyncpg==0.30.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
DateTime==5.5
email-validator==2.3.0
environs==14.1.1
fastapi==0.115.12
fastapi_cors==0.0.6
greenlet==3.2.1
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.0.0
psycopg2==2.9.11
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-httpx==0.36.0
pytest-timeout==2.4.0
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2025.2
requests==2.28.2
setuptools==79.0.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
zope.interface==7.2